Refactored JS and changed stlying
Combined multiple JS functions into one, improved consistency of the code, and ensured that the city name displayed was the data returned by the API and not the search input value. Changed the source code link colour and the border radius of the weather app. Moved Celsius and Fahrenheit links up.