Added temperature conversion functionality and styled links
Added functions to convert the current temperature, the current high, and the current low to Fahrenheit when link F is clicked, and the same functionality for converting back to Celsius. Applied CSS and JS functions to change which link is visible to the user. Corrected the spelling of Fahrenheit throughout the code!